+++ rebook-auth/src/main/java/com/be/rebook/auth/controller/AuthController.java
@@ -1,12 +1,12 @@
-package com.be.rebook.security.controller;
+package com.be.rebook.auth.controller;
 
 
-import com.be.rebook.security.entity.Members;
-import com.be.rebook.security.entity.RefreshTokens;
-import com.be.rebook.security.service.JoinService;
-import com.be.rebook.security.dto.JoinDTO;
+import com.be.rebook.auth.entity.Members;
+import com.be.rebook.auth.entity.RefreshTokens;
+import com.be.rebook.auth.service.JoinService;
+import com.be.rebook.auth.dto.JoinDTO;
 import com.be.rebook.global.config.BaseResponse;
-import com.be.rebook.security.service.ReissueService;
+import com.be.rebook.auth.service.ReissueService;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import org.slf4j.LoggerFactory;
