+++ rebook-auth/src/main/java/com/be/rebook/auth/jwt/CustomLogoutFilter.java
@@ -1,6 +1,6 @@
-package com.be.rebook.security.jwt;
+package com.be.rebook.auth.jwt;
 
-import com.be.rebook.security.repository.RefreshTokensRepository;
+import com.be.rebook.auth.repository.RefreshTokensRepository;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.ServletRequest;
