+++ rebook-auth/src/main/java/com/be/rebook/auth/repository/RefreshTokensRepository.java
@@ -1,6 +1,6 @@
-package com.be.rebook.security.repository;
+package com.be.rebook.auth.repository;
 
-import com.be.rebook.security.entity.RefreshTokens;
+import com.be.rebook.auth.entity.RefreshTokens;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
