+++ rebook-auth/src/main/java/com/be/rebook/auth/scheduler/RefreshDeleteDailyScheduler.java
@@ -1,6 +1,6 @@
-package com.be.rebook.security.scheduler;
+package com.be.rebook.auth.scheduler;
 
-import com.be.rebook.security.service.ReissueService;
+import com.be.rebook.auth.service.ReissueService;
 import org.springframework.scheduling.annotation.Scheduled;
 
 public class RefreshDeleteDailyScheduler {
