+++ rebook-auth/src/main/java/com/be/rebook/auth/service/CustomUserDetailService.java
@@ -1,10 +1,10 @@
-package com.be.rebook.security.service;
+package com.be.rebook.auth.service;
 
-import com.be.rebook.security.dto.CustomUserDetails;
+import com.be.rebook.auth.dto.CustomUserDetails;
 
 import com.be.rebook.global.exception.BaseException;
 import com.be.rebook.global.exception.ErrorCode;
-import com.be.rebook.security.repository.MembersRepository;
+import com.be.rebook.auth.repository.MembersRepository;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
